+++ altair/vegalite/v5/schema/channels.py
@@ -73,7 +73,7 @@ class ValueChannelMixin(object):
             elif 'field' in condition and 'type' not in condition:
                 kwds = parse_shorthand(condition['field'], context.get('data', None))
                 copy = self.copy(deep=['condition'])
-                copy.condition.update(kwds)
+                copy['condition'].update(kwds)
         return super(ValueChannelMixin, copy).to_dict(validate=validate,
                                                       ignore=ignore,
                                                       context=context)
@@ -97,10 +97,8 @@ class Angle(FieldChannelMixin, core.FieldOrDatumDefWithConditionMarkPropFieldDef
     """Angle schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -436,10 +434,8 @@ class AngleDatum(DatumChannelMixin, core.FieldOrDatumDefWithConditionDatumDefnum
     """AngleDatum schema wrapper
 
     Mapping(required=[])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -583,7 +579,7 @@ class AngleValue(ValueChannelMixin, core.ValueDefWithConditionMarkPropFieldOrDat
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalMarkPropFieldOrDatumDef`, :class:`ConditionalValueDefnumberExprRef`, List(:class:`ConditionalValueDefnumberExprRef`))
@@ -618,10 +614,8 @@ class Color(FieldChannelMixin, core.FieldOrDatumDefWithConditionMarkPropFieldDef
     """Color schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -957,10 +951,8 @@ class ColorDatum(DatumChannelMixin, core.FieldOrDatumDefWithConditionDatumDefGra
     """ColorDatum schema wrapper
 
     Mapping(required=[])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -1104,7 +1096,7 @@ class ColorValue(ValueChannelMixin, core.ValueDefWithConditionMarkPropFieldOrDat
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalMarkPropFieldOrDatumDef`, :class:`ConditionalValueDefGradientstringnullExprRef`, List(:class:`ConditionalValueDefGradientstringnullExprRef`))
@@ -1140,7 +1132,7 @@ class Column(FieldChannelMixin, core.RowColumnEncodingFieldDef):
 
     Mapping(required=[shorthand])
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -1452,10 +1444,8 @@ class Description(FieldChannelMixin, core.StringFieldDefWithCondition):
     """Description schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -1744,7 +1734,7 @@ class DescriptionValue(ValueChannelMixin, core.StringValueDefWithCondition):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalMarkPropFieldOrDatumDef`, :class:`ConditionalValueDefstringnullExprRef`, List(:class:`ConditionalValueDefstringnullExprRef`))
@@ -1781,7 +1771,7 @@ class Detail(FieldChannelMixin, core.FieldDefWithoutScale):
     Mapping(required=[shorthand])
     Definition object for a data field, its type and transformation of an encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -2008,7 +1998,7 @@ class Facet(FieldChannelMixin, core.FacetEncodingFieldDef):
 
     Mapping(required=[shorthand])
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -2379,10 +2369,8 @@ class Fill(FieldChannelMixin, core.FieldOrDatumDefWithConditionMarkPropFieldDefG
     """Fill schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -2718,10 +2706,8 @@ class FillDatum(DatumChannelMixin, core.FieldOrDatumDefWithConditionDatumDefGrad
     """FillDatum schema wrapper
 
     Mapping(required=[])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -2865,7 +2851,7 @@ class FillValue(ValueChannelMixin, core.ValueDefWithConditionMarkPropFieldOrDatu
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalMarkPropFieldOrDatumDef`, :class:`ConditionalValueDefGradientstringnullExprRef`, List(:class:`ConditionalValueDefGradientstringnullExprRef`))
@@ -2900,10 +2886,8 @@ class FillOpacity(FieldChannelMixin, core.FieldOrDatumDefWithConditionMarkPropFi
     """FillOpacity schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -3239,10 +3223,8 @@ class FillOpacityDatum(DatumChannelMixin, core.FieldOrDatumDefWithConditionDatum
     """FillOpacityDatum schema wrapper
 
     Mapping(required=[])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -3386,7 +3368,7 @@ class FillOpacityValue(ValueChannelMixin, core.ValueDefWithConditionMarkPropFiel
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalMarkPropFieldOrDatumDef`, :class:`ConditionalValueDefnumberExprRef`, List(:class:`ConditionalValueDefnumberExprRef`))
@@ -3421,10 +3403,8 @@ class Href(FieldChannelMixin, core.StringFieldDefWithCondition):
     """Href schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -3713,7 +3693,7 @@ class HrefValue(ValueChannelMixin, core.StringValueDefWithCondition):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalMarkPropFieldOrDatumDef`, :class:`ConditionalValueDefstringnullExprRef`, List(:class:`ConditionalValueDefstringnullExprRef`))
@@ -3750,7 +3730,7 @@ class Key(FieldChannelMixin, core.FieldDefWithoutScale):
     Mapping(required=[shorthand])
     Definition object for a data field, its type and transformation of an encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -3977,7 +3957,7 @@ class Latitude(FieldChannelMixin, core.LatLongFieldDef):
 
     Mapping(required=[shorthand])
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -4191,7 +4171,7 @@ class LatitudeDatum(DatumChannelMixin, core.DatumDef):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -4321,7 +4301,7 @@ class Latitude2(FieldChannelMixin, core.SecondaryFieldDef):
     A field definition of a secondary channel that shares a scale with another primary channel.
     For example, ``x2``, ``xError`` and ``xError2`` share the same scale with ``x``.
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -4463,7 +4443,7 @@ class Latitude2Datum(DatumChannelMixin, core.DatumDef):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -4593,7 +4573,7 @@ class Latitude2Value(ValueChannelMixin, core.PositionValueDef):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : anyOf(float, string, string, :class:`ExprRef`)
@@ -4616,7 +4596,7 @@ class Longitude(FieldChannelMixin, core.LatLongFieldDef):
 
     Mapping(required=[shorthand])
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -4830,7 +4810,7 @@ class LongitudeDatum(DatumChannelMixin, core.DatumDef):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -4960,7 +4940,7 @@ class Longitude2(FieldChannelMixin, core.SecondaryFieldDef):
     A field definition of a secondary channel that shares a scale with another primary channel.
     For example, ``x2``, ``xError`` and ``xError2`` share the same scale with ``x``.
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -5102,7 +5082,7 @@ class Longitude2Datum(DatumChannelMixin, core.DatumDef):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -5232,7 +5212,7 @@ class Longitude2Value(ValueChannelMixin, core.PositionValueDef):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : anyOf(float, string, string, :class:`ExprRef`)
@@ -5254,10 +5234,8 @@ class Opacity(FieldChannelMixin, core.FieldOrDatumDefWithConditionMarkPropFieldD
     """Opacity schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -5593,10 +5571,8 @@ class OpacityDatum(DatumChannelMixin, core.FieldOrDatumDefWithConditionDatumDefn
     """OpacityDatum schema wrapper
 
     Mapping(required=[])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -5740,7 +5716,7 @@ class OpacityValue(ValueChannelMixin, core.ValueDefWithConditionMarkPropFieldOrD
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalMarkPropFieldOrDatumDef`, :class:`ConditionalValueDefnumberExprRef`, List(:class:`ConditionalValueDefnumberExprRef`))
@@ -5776,7 +5752,7 @@ class Order(FieldChannelMixin, core.OrderFieldDef):
 
     Mapping(required=[shorthand])
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -6009,7 +5985,7 @@ class OrderValue(ValueChannelMixin, core.OrderValueDef):
 
     Mapping(required=[value])
 
-    Attributes
+    Parameters
     ----------
 
     value : anyOf(float, :class:`ExprRef`)
@@ -6046,7 +6022,7 @@ class Radius(FieldChannelMixin, core.PositionFieldDefBase):
 
     Mapping(required=[shorthand])
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -6397,7 +6373,7 @@ class RadiusDatum(DatumChannelMixin, core.PositionDatumDefBase):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -6591,7 +6567,7 @@ class RadiusValue(ValueChannelMixin, core.PositionValueDef):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : anyOf(float, string, string, :class:`ExprRef`)
@@ -6616,7 +6592,7 @@ class Radius2(FieldChannelMixin, core.SecondaryFieldDef):
     A field definition of a secondary channel that shares a scale with another primary channel.
     For example, ``x2``, ``xError`` and ``xError2`` share the same scale with ``x``.
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -6758,7 +6734,7 @@ class Radius2Datum(DatumChannelMixin, core.DatumDef):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -6888,7 +6864,7 @@ class Radius2Value(ValueChannelMixin, core.PositionValueDef):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : anyOf(float, string, string, :class:`ExprRef`)
@@ -6911,7 +6887,7 @@ class Row(FieldChannelMixin, core.RowColumnEncodingFieldDef):
 
     Mapping(required=[shorthand])
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -7223,10 +7199,8 @@ class Shape(FieldChannelMixin, core.FieldOrDatumDefWithConditionMarkPropFieldDef
     """Shape schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -7562,10 +7536,8 @@ class ShapeDatum(DatumChannelMixin, core.FieldOrDatumDefWithConditionDatumDefstr
     """ShapeDatum schema wrapper
 
     Mapping(required=[])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -7709,7 +7681,7 @@ class ShapeValue(ValueChannelMixin, core.ValueDefWithConditionMarkPropFieldOrDat
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalMarkPropFieldOrDatumDefTypeForShape`, :class:`ConditionalValueDefstringnullExprRef`, List(:class:`ConditionalValueDefstringnullExprRef`))
@@ -7744,10 +7716,8 @@ class Size(FieldChannelMixin, core.FieldOrDatumDefWithConditionMarkPropFieldDefn
     """Size schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -8083,10 +8053,8 @@ class SizeDatum(DatumChannelMixin, core.FieldOrDatumDefWithConditionDatumDefnumb
     """SizeDatum schema wrapper
 
     Mapping(required=[])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -8230,7 +8198,7 @@ class SizeValue(ValueChannelMixin, core.ValueDefWithConditionMarkPropFieldOrDatu
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalMarkPropFieldOrDatumDef`, :class:`ConditionalValueDefnumberExprRef`, List(:class:`ConditionalValueDefnumberExprRef`))
@@ -8265,10 +8233,8 @@ class Stroke(FieldChannelMixin, core.FieldOrDatumDefWithConditionMarkPropFieldDe
     """Stroke schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -8604,10 +8570,8 @@ class StrokeDatum(DatumChannelMixin, core.FieldOrDatumDefWithConditionDatumDefGr
     """StrokeDatum schema wrapper
 
     Mapping(required=[])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -8751,7 +8715,7 @@ class StrokeValue(ValueChannelMixin, core.ValueDefWithConditionMarkPropFieldOrDa
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalMarkPropFieldOrDatumDef`, :class:`ConditionalValueDefGradientstringnullExprRef`, List(:class:`ConditionalValueDefGradientstringnullExprRef`))
@@ -8786,10 +8750,8 @@ class StrokeDash(FieldChannelMixin, core.FieldOrDatumDefWithConditionMarkPropFie
     """StrokeDash schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -9125,10 +9087,8 @@ class StrokeDashDatum(DatumChannelMixin, core.FieldOrDatumDefWithConditionDatumD
     """StrokeDashDatum schema wrapper
 
     Mapping(required=[])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -9272,7 +9232,7 @@ class StrokeDashValue(ValueChannelMixin, core.ValueDefWithConditionMarkPropField
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalMarkPropFieldOrDatumDef`, :class:`ConditionalValueDefnumberArrayExprRef`, List(:class:`ConditionalValueDefnumberArrayExprRef`))
@@ -9307,10 +9267,8 @@ class StrokeOpacity(FieldChannelMixin, core.FieldOrDatumDefWithConditionMarkProp
     """StrokeOpacity schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -9646,10 +9604,8 @@ class StrokeOpacityDatum(DatumChannelMixin, core.FieldOrDatumDefWithConditionDat
     """StrokeOpacityDatum schema wrapper
 
     Mapping(required=[])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -9793,7 +9749,7 @@ class StrokeOpacityValue(ValueChannelMixin, core.ValueDefWithConditionMarkPropFi
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalMarkPropFieldOrDatumDef`, :class:`ConditionalValueDefnumberExprRef`, List(:class:`ConditionalValueDefnumberExprRef`))
@@ -9828,10 +9784,8 @@ class StrokeWidth(FieldChannelMixin, core.FieldOrDatumDefWithConditionMarkPropFi
     """StrokeWidth schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -10167,10 +10121,8 @@ class StrokeWidthDatum(DatumChannelMixin, core.FieldOrDatumDefWithConditionDatum
     """StrokeWidthDatum schema wrapper
 
     Mapping(required=[])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -10314,7 +10266,7 @@ class StrokeWidthValue(ValueChannelMixin, core.ValueDefWithConditionMarkPropFiel
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalMarkPropFieldOrDatumDef`, :class:`ConditionalValueDefnumberExprRef`, List(:class:`ConditionalValueDefnumberExprRef`))
@@ -10349,10 +10301,8 @@ class Text(FieldChannelMixin, core.FieldOrDatumDefWithConditionStringFieldDefTex
     """Text schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -10640,10 +10590,8 @@ class TextDatum(DatumChannelMixin, core.FieldOrDatumDefWithConditionStringDatumD
     """TextDatum schema wrapper
 
     Mapping(required=[])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -10833,7 +10781,7 @@ class TextValue(ValueChannelMixin, core.ValueDefWithConditionStringFieldDefText)
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalStringFieldDef`, :class:`ConditionalValueDefTextExprRef`, List(:class:`ConditionalValueDefTextExprRef`))
@@ -10869,7 +10817,7 @@ class Theta(FieldChannelMixin, core.PositionFieldDefBase):
 
     Mapping(required=[shorthand])
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -11219,7 +11167,7 @@ class ThetaDatum(DatumChannelMixin, core.PositionDatumDefBase):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -11413,7 +11361,7 @@ class ThetaValue(ValueChannelMixin, core.PositionValueDef):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : anyOf(float, string, string, :class:`ExprRef`)
@@ -11438,7 +11386,7 @@ class Theta2(FieldChannelMixin, core.SecondaryFieldDef):
     A field definition of a secondary channel that shares a scale with another primary channel.
     For example, ``x2``, ``xError`` and ``xError2`` share the same scale with ``x``.
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -11580,7 +11528,7 @@ class Theta2Datum(DatumChannelMixin, core.DatumDef):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -11710,7 +11658,7 @@ class Theta2Value(ValueChannelMixin, core.PositionValueDef):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : anyOf(float, string, string, :class:`ExprRef`)
@@ -11732,10 +11680,8 @@ class Tooltip(FieldChannelMixin, core.StringFieldDefWithCondition):
     """Tooltip schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -12024,7 +11970,7 @@ class TooltipValue(ValueChannelMixin, core.StringValueDefWithCondition):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalMarkPropFieldOrDatumDef`, :class:`ConditionalValueDefstringnullExprRef`, List(:class:`ConditionalValueDefstringnullExprRef`))
@@ -12059,10 +12005,8 @@ class Url(FieldChannelMixin, core.StringFieldDefWithCondition):
     """Url schema wrapper
 
     Mapping(required=[shorthand])
-    A FieldDef with Condition :raw-html:`<ValueDef>` {   condition: {value: ...},   field: ...,
-    ... }
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -12351,7 +12295,7 @@ class UrlValue(ValueChannelMixin, core.StringValueDefWithCondition):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     condition : anyOf(:class:`ConditionalMarkPropFieldOrDatumDef`, :class:`ConditionalValueDefstringnullExprRef`, List(:class:`ConditionalValueDefstringnullExprRef`))
@@ -12387,7 +12331,7 @@ class X(FieldChannelMixin, core.PositionFieldDef):
 
     Mapping(required=[shorthand])
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -12771,7 +12715,7 @@ class XDatum(DatumChannelMixin, core.PositionDatumDef):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     axis : anyOf(:class:`Axis`, None)
@@ -12998,7 +12942,7 @@ class XValue(ValueChannelMixin, core.PositionValueDef):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : anyOf(float, string, string, :class:`ExprRef`)
@@ -13023,7 +12967,7 @@ class X2(FieldChannelMixin, core.SecondaryFieldDef):
     A field definition of a secondary channel that shares a scale with another primary channel.
     For example, ``x2``, ``xError`` and ``xError2`` share the same scale with ``x``.
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -13164,7 +13108,7 @@ class X2Datum(DatumChannelMixin, core.DatumDef):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -13294,7 +13238,7 @@ class X2Value(ValueChannelMixin, core.PositionValueDef):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : anyOf(float, string, string, :class:`ExprRef`)
@@ -13319,7 +13263,7 @@ class XError(FieldChannelMixin, core.SecondaryFieldDef):
     A field definition of a secondary channel that shares a scale with another primary channel.
     For example, ``x2``, ``xError`` and ``xError2`` share the same scale with ``x``.
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -13463,7 +13407,7 @@ class XErrorValue(ValueChannelMixin, core.ValueDefnumber):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : float
@@ -13488,7 +13432,7 @@ class XError2(FieldChannelMixin, core.SecondaryFieldDef):
     A field definition of a secondary channel that shares a scale with another primary channel.
     For example, ``x2``, ``xError`` and ``xError2`` share the same scale with ``x``.
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -13632,7 +13576,7 @@ class XError2Value(ValueChannelMixin, core.ValueDefnumber):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : float
@@ -13655,7 +13599,7 @@ class XOffset(FieldChannelMixin, core.ScaleFieldDef):
 
     Mapping(required=[shorthand])
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -13959,7 +13903,7 @@ class XOffsetDatum(DatumChannelMixin, core.ScaleDatumDef):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -14111,7 +14055,7 @@ class XOffsetValue(ValueChannelMixin, core.ValueDefnumber):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : float
@@ -14134,7 +14078,7 @@ class Y(FieldChannelMixin, core.PositionFieldDef):
 
     Mapping(required=[shorthand])
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -14518,7 +14462,7 @@ class YDatum(DatumChannelMixin, core.PositionDatumDef):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     axis : anyOf(:class:`Axis`, None)
@@ -14745,7 +14689,7 @@ class YValue(ValueChannelMixin, core.PositionValueDef):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : anyOf(float, string, string, :class:`ExprRef`)
@@ -14770,7 +14714,7 @@ class Y2(FieldChannelMixin, core.SecondaryFieldDef):
     A field definition of a secondary channel that shares a scale with another primary channel.
     For example, ``x2``, ``xError`` and ``xError2`` share the same scale with ``x``.
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -14911,7 +14855,7 @@ class Y2Datum(DatumChannelMixin, core.DatumDef):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -15041,7 +14985,7 @@ class Y2Value(ValueChannelMixin, core.PositionValueDef):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : anyOf(float, string, string, :class:`ExprRef`)
@@ -15066,7 +15010,7 @@ class YError(FieldChannelMixin, core.SecondaryFieldDef):
     A field definition of a secondary channel that shares a scale with another primary channel.
     For example, ``x2``, ``xError`` and ``xError2`` share the same scale with ``x``.
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -15210,7 +15154,7 @@ class YErrorValue(ValueChannelMixin, core.ValueDefnumber):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : float
@@ -15235,7 +15179,7 @@ class YError2(FieldChannelMixin, core.SecondaryFieldDef):
     A field definition of a secondary channel that shares a scale with another primary channel.
     For example, ``x2``, ``xError`` and ``xError2`` share the same scale with ``x``.
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -15379,7 +15323,7 @@ class YError2Value(ValueChannelMixin, core.ValueDefnumber):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : float
@@ -15402,7 +15346,7 @@ class YOffset(FieldChannelMixin, core.ScaleFieldDef):
 
     Mapping(required=[shorthand])
 
-    Attributes
+    Parameters
     ----------
 
     shorthand : string
@@ -15706,7 +15650,7 @@ class YOffsetDatum(DatumChannelMixin, core.ScaleDatumDef):
 
     Mapping(required=[])
 
-    Attributes
+    Parameters
     ----------
 
     bandPosition : float
@@ -15858,7 +15802,7 @@ class YOffsetValue(ValueChannelMixin, core.ValueDefnumber):
     Definition object for a constant value (primitive value or gradient definition) of an
     encoding channel.
 
-    Attributes
+    Parameters
     ----------
 
     value : float